+++ src/main/java/com/bokmcdok/butterflies/event/entity/EntityEventListener.java
@@ -1,19 +1,11 @@
 package com.bokmcdok.butterflies.event.entity;
 
-import com.bokmcdok.butterflies.ButterfliesMod;
 import com.bokmcdok.butterflies.registries.EntityTypeRegistry;
 import com.bokmcdok.butterflies.world.entity.animal.*;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Mob;
-import net.minecraft.world.entity.SpawnPlacements;
+import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.ai.goal.target.NearestAttackableTargetGoal;
 import net.minecraft.world.entity.ai.goal.target.NonTameRandomTargetGoal;
-import net.minecraft.world.entity.animal.Cat;
-import net.minecraft.world.entity.animal.Fox;
-import net.minecraft.world.entity.animal.Ocelot;
-import net.minecraft.world.entity.animal.Parrot;
-import net.minecraft.world.entity.animal.Wolf;
+import net.minecraft.world.entity.animal.*;
 import net.minecraft.world.entity.monster.Spider;
 import net.minecraft.world.entity.monster.Witch;
 import net.minecraft.world.entity.monster.Zombie;
@@ -42,7 +34,7 @@ public class EntityEventListener {
                                IEventBus modEventBus,
                                EntityTypeRegistry entityTypeRegistry) {
         forgeEventBus.register(this);
-        forgeEventBus.addListener(this::onJoinWorld);
+        forgeEventBus.addListener(this::onEntityJoinLevel);
 
         modEventBus.register(this);
         modEventBus.addListener(this::onEntityAttributeCreation);
@@ -83,13 +75,15 @@ public class EntityEventListener {
         for (RegistryObject<EntityType<ButterflyEgg>> i : entityTypeRegistry.getButterflyEggs()) {
             event.put(i.get(), ButterflyEgg.createAttributes().build());
         }
+
+        event.put(entityTypeRegistry.getButterflyGolem().get(), IronGolem.createAttributes().build());
     }
 
     /**
      * On joining the world modify entities' goals so butterflies have predators.
      * @param event Information for the event.
      */
-    private void onJoinWorld(EntityJoinLevelEvent event) {
+    private void onEntityJoinLevel(EntityJoinLevelEvent event) {
 
         //  Cat
         if (event.getEntity() instanceof Cat cat) {